confirm dialog primeng

+++ src/app/transaction/transaction-search/transaction-search.component.ts
@@ -4,12 +4,12 @@ import { LazyLoadEvent } from 'primeng/components/common/lazyloadevent';
 
 import { TransactionService, TransactionFilter } from './../transaction.service';
 import { MessageService } from 'primeng/components/common/messageservice';
+import { ConfirmationService } from 'primeng/components/common/confirmationservice';
 
 @Component({
   selector: 'app-transaction-search',
   templateUrl: './transaction-search.component.html',
-  styleUrls: ['./transaction-search.component.css'],
-  providers: [MessageService]
+  styleUrls: ['./transaction-search.component.css']
 })
 export class TransactionSearchComponent implements OnInit {
 
@@ -20,7 +20,8 @@ export class TransactionSearchComponent implements OnInit {
 
   constructor(
     private transactionService: TransactionService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ){}
 
   ngOnInit() {
@@ -42,15 +43,22 @@ export class TransactionSearchComponent implements OnInit {
   }
 
   delete(transaction : any){
-    this.transactionService.delete(transaction.id)
-      .subscribe(() => {
-        if(this.grid.first === 0){
-          this.findAllSummary();
-        }else{
-          this.grid.first = 0;
-        }
-        --this.totalElements;
-        this.messageService.add({severity: 'success', summary: 'Success', detail: 'Transaction has been deleted'})
-      })
+    this.confirmationService.confirm({
+      message: 'Do you want to delete this record?',
+      header: 'Delete Confirmation',
+      icon: 'fa fa-trash',
+      accept: () => {
+        this.transactionService.delete(transaction.id)
+        .subscribe(() => {
+          if(this.grid.first === 0){
+            this.findAllSummary();
+          }else{
+            this.grid.first = 0;
+          }
+          --this.totalElements;
+          this.messageService.add({severity: 'success', summary: 'Success', detail: 'Transaction has been deleted'})
+        })
+      }
+    });
   }
 }
